Refactor to have two different methods to request and train
