+++ demo-app/src/main/java/org/openmined/syft/demo/federated/ui/main/MnistActivityViewModel.kt
@@ -14,8 +14,8 @@ import org.openmined.syft.demo.federated.service.LOG
 import org.openmined.syft.demo.federated.service.LOSS_LIST
 import org.openmined.syft.demo.federated.service.STATUS
 import org.openmined.syft.demo.federated.service.WorkerRepository
-import org.openmined.syft.demo.federated.ui.ContentState
-import org.openmined.syft.demo.federated.ui.ProcessData
+import org.openmined.syft.domain.ContentState
+import org.openmined.syft.domain.ProcessData
 import org.openmined.syft.domain.SyftConfiguration
 
 @ExperimentalUnsignedTypes
