+++ syft/src/test/java/org/openmined/syft/integration/AuthenticationTest.kt
@@ -51,7 +51,7 @@ class AuthenticationTest : AbstractSyftWorkerTest() {
         val syftWorker = Syft.getInstance(syftConfiguration, "test token")
         val job = syftWorker.newJob("test", "1")
         val jobStatusSubscriber = spy<JobStatusSubscriber>()
-        job.start(jobStatusSubscriber)
+        job.request(jobStatusSubscriber)
         verify(jobStatusSubscriber).onReady(any(), any(), any())
         syftWorker.dispose()
         verify(jobStatusSubscriber).onComplete()
@@ -88,7 +88,7 @@ class AuthenticationTest : AbstractSyftWorkerTest() {
         val syftWorker = Syft.getInstance(syftConfiguration)
         val job = syftWorker.newJob("test", "1")
         val jobStatusSubscriber = spy<JobStatusSubscriber>()
-        job.start(jobStatusSubscriber)
+        job.request(jobStatusSubscriber)
         verify(jobStatusSubscriber).onReady(any(), any(), any())
         syftWorker.dispose()
         verify(jobStatusSubscriber).onComplete()
@@ -125,7 +125,7 @@ class AuthenticationTest : AbstractSyftWorkerTest() {
         val syftWorker = Syft.getInstance(syftConfiguration)
         val job = syftWorker.newJob("test", "1")
         val jobStatusSubscriber = spy<JobStatusSubscriber>()
-        job.start(jobStatusSubscriber)
+        job.request(jobStatusSubscriber)
         argumentCaptor<Throwable>().apply {
             verify(jobStatusSubscriber).onError(capture())
             assert(firstValue is JobErrorThrowable.AuthenticationFailure)
@@ -164,7 +164,7 @@ class AuthenticationTest : AbstractSyftWorkerTest() {
         val syftWorker = Syft.getInstance(syftConfiguration)
         val job = syftWorker.newJob("test", "1")
         val jobStatusSubscriber = spy<JobStatusSubscriber>()
-        job.start(jobStatusSubscriber)
+        job.request(jobStatusSubscriber)
         verify(jobStatusSubscriber).onRejected(any())
         syftWorker.dispose()
         verify(jobStatusSubscriber).onComplete()
