+++ syft/src/test/java/org/openmined/syft/integration/DownloadablesTest.kt
@@ -49,7 +49,7 @@ class DownloadablesTest : AbstractSyftWorkerTest() {
         val syftWorker = Syft.getInstance(syftConfiguration)
         val job = syftWorker.newJob("test", "1")
         val jobStatusSubscriber = spy<JobStatusSubscriber>()
-        job.start(jobStatusSubscriber)
+        job.request(jobStatusSubscriber)
         verify(jobStatusSubscriber).onError(any())
         syftWorker.dispose()
         verify(jobStatusSubscriber, never()).onComplete()
@@ -86,7 +86,7 @@ class DownloadablesTest : AbstractSyftWorkerTest() {
         val syftWorker = Syft.getInstance(syftConfiguration)
         val job = syftWorker.newJob("test", "1")
         val jobStatusSubscriber = spy<JobStatusSubscriber>()
-        job.start(jobStatusSubscriber)
+        job.request(jobStatusSubscriber)
         verify(jobStatusSubscriber).onError(any())
         syftWorker.dispose()
         verify(jobStatusSubscriber, never()).onComplete()
