+++ syft/src/test/java/org/openmined/syft/integration/SpeedIntegrationTest.kt
@@ -49,7 +49,7 @@ class SpeedIntegrationTest : AbstractSyftWorkerTest() {
         val syftWorker = Syft.getInstance(syftConfiguration)
         val job = syftWorker.newJob("test", "1")
         val jobStatusSubscriber = spy<JobStatusSubscriber>()
-        job.start(jobStatusSubscriber)
+        job.request(jobStatusSubscriber)
         verify(jobStatusSubscriber).onRejected(any())
         syftWorker.dispose()
         verify(jobStatusSubscriber).onComplete()
@@ -86,7 +86,7 @@ class SpeedIntegrationTest : AbstractSyftWorkerTest() {
         val syftWorker = Syft.getInstance(syftConfiguration)
         val job = syftWorker.newJob("test", "1")
         val jobStatusSubscriber = spy<JobStatusSubscriber>()
-        job.start(jobStatusSubscriber)
+        job.request(jobStatusSubscriber)
         verify(jobStatusSubscriber).onError(any())
         syftWorker.dispose()
         verify(jobStatusSubscriber, never()).onComplete()
@@ -123,7 +123,7 @@ class SpeedIntegrationTest : AbstractSyftWorkerTest() {
         val syftWorker = Syft.getInstance(syftConfiguration)
         val job = syftWorker.newJob("test", "1")
         val jobStatusSubscriber = spy<JobStatusSubscriber>()
-        job.start(jobStatusSubscriber)
+        job.request(jobStatusSubscriber)
         verify(jobStatusSubscriber).onError(any())
         syftWorker.dispose()
         verify(jobStatusSubscriber, never()).onComplete()
@@ -160,7 +160,7 @@ class SpeedIntegrationTest : AbstractSyftWorkerTest() {
         val syftWorker = Syft.getInstance(syftConfiguration)
         val job = syftWorker.newJob("test", "1")
         val jobStatusSubscriber = spy<JobStatusSubscriber>()
-        job.start(jobStatusSubscriber)
+        job.request(jobStatusSubscriber)
         verify(jobStatusSubscriber).onError(any())
         syftWorker.dispose()
         verify(jobStatusSubscriber, never()).onComplete()
